==> Scripts/02-Analyze.Rmd <==
---
title: "02-Analyze"
author: "Ian Adams"
date: "3/27/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      message = FALSE,
                      warning = FALSE)

library(tidyverse)
library(readr)
library(here)

#library(ggplot2)
#library(lubridate)

library(gghighlight)
library(ggstance)
library(ggalt)

library(ggthemes)
library(extrafont)
library(ggrepel)

library(readxl)
#library(stringr)

# Load fonts
loadfonts(device = "win", quiet = TRUE)


```

## Source new data

This is just running the source file `01-Import V2.R` which pulls the latest WaPo data, merges it, and saves it out.

```{r setup, include=FALSE}

# Source the Wapo data import .R script

#source("~/Library/CloudStorage/GoogleDrive-adams.po@gmail.com/Other computers/My Computer/R/WAPO data-police-shootings/Scripts/01-Import V2.R")

source("Scripts/01-Import V2.R")

df_all$date <- ymd(df_all$date) # format as lubridate var
df_all$year <- lubridate::year(df_all$date) # extract year and create variable year
df_all$month <- month(df_all$date, label = TRUE, abbr = TRUE) # extract month and create variable month
df_all$day <- wday(df_all$date, label = TRUE, abbr = TRUE) # extract day of week, create variable day

```

## Cumulative line plots like jnixy's

```{r remedy007}

# Load the required packages
library(ggplot2)
library(dplyr)
library(ggrepel)
library(ggthemes)
library(extrafont)

# Assuming 'df_all' is your dataset and 'date' is the column storing the shooting dates
df <- df_all

# Shape data
shootings_data <- df %>%
  arrange(date) %>%
  mutate(year = as.numeric(format(date, "%Y"))) %>%
  group_by(year, date) %>%
  summarize(daily_count = n(), .groups = 'drop') %>%
  ungroup()

shootings_data <- shootings_data %>%
  group_by(year) %>%
  mutate(cum_count = cumsum(daily_count)) %>%
  ungroup()

# Load fonts (only if necessary, usually on Windows)
# loadfonts(device = "win", quiet = TRUE)

# Store the current date in a variable
current_date <- Sys.Date()
formatted_date <- format(current_date, "%B %d, %Y")

# Determine the current year
current_year <- as.numeric(format(Sys.Date(), "%Y"))

# Calculate current year's shootings count (ensure 2024 data is included in df_all)
current_year_count <- shootings_data %>%
  filter(year == current_year) %>%
  summarize(total_count = max(cum_count)) %>%
  pull(total_count)

# Determine the first and latest year in the data
min_year <- min(shootings_data$year)
latest_year <- max(shootings_data$year)

# Dynamic title construction
plot_title <- paste("Cumulative U.S. Police Lethal Shootings by Year,", min_year, "-", latest_year)

# Plot the figure
p <- ggplot(shootings_data, aes(x = format(date, "%m-%d"), y = cum_count, group = factor(year))) +
  geom_point(aes(color = factor(year)), alpha = 0.5, size = 1) +
  gghighlight(year == latest_year, use_direct_label = FALSE) +
  scale_x_discrete(name = "Month",
                   breaks = format(seq(as.Date("2022-01-01"), as.Date("2022-12-31"), by = "1 month"), "%m-%d"),
                   labels = format(seq(as.Date("2022-01-01"), as.Date("2022-12-31"), by = "1 month"), "%b"),
                   expand = c(0.01, 0.9)) +
  scale_y_continuous(name = "Cumulative Count of Deaths", limits = c(0, 1150)) +
  theme_tufte() +
  theme(panel.grid.minor = element_blank(),
        legend.position = "none",
        text = element_text(family = "Arial", color = "black"),
        axis.text.x = element_text(angle = 90, vjust = 0.5),
        axis.text.y = element_text(angle = 0, hjust = 0.5),
        plot.title = element_text(size = 14, hjust = 0, face = "bold"),
        plot.title.position = "plot",
        plot.caption = element_text(size = 10, hjust = 1)) +
  geom_text_repel(data = shootings_data %>% group_by(year) %>% filter(date == max(date)),
                  aes(x = format(date, "%m-%d"), y = cum_count, label = paste(year, "(", cum_count, ")")),
                  size = 3) +
  labs(title = plot_title,
       subtitle = paste("as of", formatted_date),
       caption = "Source: Ian T. Adams \ndata via Washington Post")

p

# Save it
ggsave(filename = "Output/cumulative_daily_count.jpg", plot = p, height = 8, width = 8)


```

## Race Breakdowns

```{r remedy005}

## Recoding df_all$race
df_all$race <- df_all$race %>%
  fct_recode(
    "Asian" = "A",
    "Black" = "B",
    "Other" = "B;H",
    "Hispanic" = "H",
    "Native American" = "N",
    "Other" = "O",
    "White" = "W",
    "Other" = "N;H",
    "Other" = "W;B",
    "Other" = "W;B;N",
    "Other" = "W;H"
  )


#Make "NA" into an explicit factor level
df_all$race <- df_all$race %>%
  fct_na_value_to_level("Unknown")


## Percentage by race
temp <- table(df_all$race) / nrow(df_all) * 100
temp <- sort(temp)
temp <- round(temp, digits = 2)
temp

## Barplot by race

df_all$race <- factor(df_all$race,
                         levels = names(sort(table(df_all$race), 
                                             decreasing = TRUE)))
# Store the current date in a variable
current_date <- Sys.Date()
formatted_date <- format(current_date, "%B %d, %Y")

# Determine the first and latest year in the data
min_year <- min(df_all$year)
latest_year <- max(df_all$year)

# Dynamic title construction
plot_title <- paste("Distribution of Police Lethal Shootings by Race,", min_year, "-", latest_year)

ggplot(df_all, aes(x = race)) + 
  geom_bar(fill = "steelblue", color = "black") +
  geom_text(aes(label = sprintf("%0.2f%%", (..count..)/sum(..count..)*100)),
            stat = "count", vjust = -0.5, color = "black") +
  theme_tufte() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        text = element_text(family = "Arial", color = "black"),
        axis.title.x = element_blank(),
        axis.title.y = element_text(size = 12),
        axis.text.x = element_text(angle = 45, hjust = 1),
        plot.title = element_text(size = 14, hjust = 0, face = "bold"),
        plot.title.position = "plot") +
  labs(y = "Count", title = plot_title, subtitle = paste("as of", formatted_date),
       caption = "Source: Ian T. Adams \ndata via Washington Post")

# Save the plot
ggsave(filename = here("Output", "race_distribution_barplot.jpg"), height = 8, width = 8)


##########
######### Unknown race plot

# extended color palette
colorsb <- c("#0072b2","#d55e00", "#009e73", "#f0e442", "#56b4e9", "#cc79a7", "#999999", "#E69F00")


# Calculate the percentage of undetermined race for each year
df_percent <- df_all %>%
  group_by(year, race) %>%
  summarize(count = n()) %>%
  mutate(total = sum(count),
         percent = count / total * 100) %>%
  filter(race == "Unknown")

# Create the bar plot
p <- df_all %>% ggplot(aes(x=year, fill=race)) +
  geom_bar(col = 'white') +
  scale_fill_manual(values = colorsb) +
  labs(
    x = 'Year',
    y = 'Counts',
    title = 'Fatal Police Shootings by Year and Race (with % of unknown)',
    subtitle = paste0('as of ', formatted_date),
    caption = "Source: Ian T. Adams \ndata via Washington Post",
    fill = 'Race'
  ) + scale_x_continuous(breaks = c(2015,2016,2017,2018,2019,2020,2021,2022,2023, 2024)) +
  gghighlight(race == 'Unknown', use_direct_label = F)

# Add percentage labels for undetermined race
p + geom_text(data = df_percent, aes(x = year, y = count, label = paste0(round(percent, 1), "%")), position = position_stack(vjust = 0.5), color = "black")

# Save the plot
ggsave(filename = here("Output", "undetermined_highlight.png"), bg="white")

##### barplot with all races no highlight
# Create the bar plot
p <- df_all %>% ggplot(aes(x=year, fill=race)) +
  geom_bar(col = 'white') +
  scale_fill_manual(values = colorsb) +
  labs(
    x = 'Year',
    y = 'Counts',
    title = 'Fatal Police Shootings by Year and Race (with % of unknown)',
    subtitle = paste0('as of ', formatted_date),
    caption = "Source: Ian T. Adams \ndata via Washington Post",
    fill = 'Race'
  ) + scale_x_continuous(breaks = c(2015,2016,2017,2018,2019,2020,2021,2022,2023, 2024)) 

# Add percentage labels for undetermined race
p + geom_text(data = df_percent, aes(x = year, y = count, label = paste0(round(percent, 1), "%")), position = position_stack(vjust = 0.5), color = "black")

# Save the plot
ggsave(filename = here("Output", "undetermined.png"), bg="white")

############
############ Rolling average of unknowns

library(slider)

# Assuming df_all has a date column in YYYY-MM-DD format
df_all$date <- as.Date(df_all$date)

# Set the rolling window size (e.g., 60 days)
rolling_window <- 90  # Can be changed to 30, 90, etc.

# Calculate the rolling percentage of unknown race
rolling_percent_unknown <- df_all %>%
  group_by(date) %>%
  summarize(unknown_count = sum(race == "Unknown", na.rm = TRUE),
            total_count = n()) %>%
  ungroup() %>%
  mutate(unknown_percent = unknown_count / total_count * 100) %>%
  mutate(rolling_avg = slide_dbl(unknown_percent, mean, .before = rolling_window - 1, .complete = TRUE))

# Plotting
ggplot(rolling_percent_unknown, aes(x = date, y = rolling_avg)) +
  geom_line(color = "steelblue") +
  theme_minimal() +
  labs(x = "Date",
       y = "Rolling Average of Unknown Race (%)",
       title = paste("Rolling Average of Unknown Race Percentage in Police Shootings"),
       subtitle = paste("Rolling window:", rolling_window, "days"),
       caption = "Source: Ian T. Adams \ndata via Washington Post")

# Save the plot
ggsave(filename = here("Output", "rolling_avg_unknown_race.jpg"), height = 8, width = 12)








## Or flipped and with proportion

# ggplot(df_all, aes(x = race)) + 
#   geom_bar() +
#   coord_flip() +
#   geom_text(aes(label = sprintf("%0.2f", round(..count..)/sum(..count..),digits=2)), stat = "count", hjust = 0, colour = "black")

```

## National dot map simple

```{r remedy003}

# Load the packages

library(ggmap)
library(maps)
library(sf)
library("rnaturalearth")
library("rnaturalearthdata")


# Filter out rows with NA in latitude or longitude
df_filtered <- df_all %>%
    filter(!is.na(latitude) & !is.na(longitude))

world <- ne_countries(scale = "medium", returnclass = "sf")
class(world)

sites <- st_as_sf(df_filtered, coords = c("longitude", "latitude"), 
    crs = 4326, agr = "constant")

ggplot(data = world) +
    geom_sf() +
    geom_sf(data = sites, size = .5, shape = 10, fill = "darkred") +
    # You can adjust these limits to show the whole USA
    coord_sf(xlim = c(-125, -66), ylim = c(24, 49), expand = TRUE)



# Save the plot
ggsave(filename = here("Output", "state_pins_all.jpg"), height = 8, width = 12)

```

## Armed Status Breakdown 


```{r remedy002, include=FALSE}

ggplot(df, aes(x = armed_with)) +
  geom_bar(fill = "#619CFF") +
  labs(title = "Weapons used in police shootings",
       x = "Armed with",
       y = "Count") +
  theme_minimal()

ggsave(p3, filename = here("Output", "time_armed_threat.jpg"))

```


## Most Common Armed Weapons

```{r remedy003}

temp <- table(df_all$armed_with) / nrow(df) * 100
temp <- sort(temp, decreasing = TRUE)
temp <- round(temp, digits = 2)
temp
```
## Gender Breakdowns

```{r remedy004}
table(df_all$gender) / nrow(df) * 100

```

## Flee breakdowns

```{r remedy007}
table(df_all$flee_status) / nrow(df) * 100

```

## Age Breakdowns

```{r remedy006}
## Histogram
ggplot(df_all, aes(x = age)) + 
  geom_histogram(aes(y = (..count..)/sum(..count..)))

## Density

ggplot(df_all, aes(x = age)) + 
  geom_density()

## Simple Stat Count

ggplot(df_all, aes(x = age)) + 
  stat_count(aes(y = (..count..)/sum(..count..)))

```



## Over time with population

```{r remedy007}

library(tidycensus)

# set your Census API key here
census_api_key("e0ec52f1c5358231d666ff8dce8c15eb40002b78", install = TRUE, overwrite = TRUE)

#Manually create population tibble
# sourced from https://www.worldometers.info/world-population/us-population/
pop_data <- tribble(
  ~year, ~estimate,
  2015, 324607776,
  2016, 327210198,
  2017, 329791231,
  2018, 332140037,
  2019, 334319671,
  2020, 335942003,
  2021, 336997624,
  2022, 338289857,
  2023, 339996563
)

pop_data <- pop_data %>% mutate(NAME = "United States")

df2023 <- df_all %>%
  filter(year >= 2015 & year <= 2023)

# join with events data and calculate normalized events
events_normalized <- df2023 %>%
  mutate(year = lubridate::year(date)) %>%
  group_by(year) %>%
  summarize(total_events = n()) %>%
  left_join(pop_data, by = "year") %>%
  mutate(events_per_capita = total_events / estimate * 1000000)

ggplot(events_normalized, aes(x = year, y = events_per_capita)) +
  geom_bar(stat = "identity") +
  labs(title = "US Police Shooting Deaths per 1,000,000 people by year") +
  labs(y="Fatal shootings per capita") +
  labs(caption = "Source: Washington Post & United Nations - World Population Prospects")

```

## Add uncertainty

```{r remedy008}


# Define color palette for the bars
colors <- c("#d3d3d3", "#ff7f0e", "#1f77b4", "#2ca02c", "#9467bd", "#8c564b", "#e377c2", "#7f7f7f")

# Join with events data and calculate normalized events
events_normalized <- df2023 %>%
  mutate(year = lubridate::year(date)) %>%
  group_by(year) %>%
  summarize(total_events = n()) %>%
  left_join(pop_data, by = "year") %>%
  mutate(
    events_per_capita = total_events / estimate * 1000000,
    se = events_per_capita / sqrt(0.92 * total_events)
  )

# Plot the bar chart with error bars and labels
ggplot(events_normalized, aes(x = year, y = events_per_capita)) +
  geom_bar(stat = "identity", color = "white", size = 0.2, fill="#7777AB") +
  #scale_fill_manual(values = colors) +
  geom_errorbar(aes(ymin = events_per_capita, ymax = events_per_capita + 1.96 * se), 
                 width = 0.2, color = "black", size = 0.8) +
  geom_text(aes(label = sprintf("%.2f", events_per_capita), y = events_per_capita/2), vjust = 1.5) +
  labs(title = "US Police Shooting Deaths per 1,000,000 people by year") +
  labs(subtitle = "assuming 92% counting accuracy") +
  labs(y = "Fatal shootings per capita") +
  labs(x = "Year") +
  labs(caption = "Source: Washington Post and United Nations - World Population Prospects \n Courtesy: Ian Adams and Justin Nix") +
  theme_bw() +
  theme(
    axis.text = element_text(size = 12),
    axis.title = element_text(size = 14),
    plot.title = element_text(size = 16, face = "bold"),
    legend.position = "none"
  ) +
  scale_x_continuous(breaks = events_normalized$year, labels = events_normalized$year)


ggsave(filename = here("Output", "pop_normal_error.jpg"), height = 5, width = 8)

```

## Top 15 States

```{r, fig.height=9, fig.width=9}
stateinfo <- df_all %>% group_by(state) %>% summarise(n = n()) %>% 
        arrange(desc(n)) %>% top_n(15) %>% 
        mutate(state = factor(state, levels = rev(unique(state))))

ggplot(stateinfo) +
  aes(x = state, fill = state, weight = n) +
  geom_bar() +
  scale_fill_manual(
    values = c(PA = "#A6CEE3",
    LA = "#3B8ABE",
    NM = "#72B29C",
    MO = "#84C868",
    OH = "#4F9F3B",
    WA = "#EC9A91",
    TN = "#E93E3F",
    OK = "#F06C45",
    NC = "#FDAC4F",
    CO = "#FB820F",
    GA = "#D1AAB7",
    AZ = "#8C66AF",
    FL = "#A99099",
    TX = "#EEDB80",
    CA = "#B15928")
  ) +
  labs(
    y = "Number of deaths",
    title = "Top 15 States for Police Fatal Shootings",
    subtitle = "2015-2022",
    caption = "Washington Post data"
  ) +
  coord_flip() +
  hrbrthemes::theme_ipsum_es() +
  theme(legend.position = "none")
```


## Armed Status? (top 15 reported categories)

```{r}

df_all$armed_with <- df_all$armed_with %>%
  fct_explicit_na("unknown")

armedinfo <- df_all %>% group_by(armed_with) %>% summarise(n = n()) %>% 
        arrange(desc(n)) %>% top_n(15) %>% 
        mutate(armed = factor(armed_with, levels = rev(unique(armed_with))))

## Recoding armedinfo$armed_with into armedinfo$armed_with_rec
armedinfo$armed_with_rec <- armedinfo$armed_with %>%
  fct_recode(
    "Blunt" = "blunt_object",
    "Blunt" = "blunt_object;blunt_object",
    "Knife" = "blunt_object;knife",
    "Blunt" = "blunt_object;other",
    "Gun" = "gun",
    "Gun" = "gun;knife",
    "Gun" = "gun;other",
    "Gun" = "gun;vehicle",
    "Knife" = "knife",
    "Knife" = "knife;blunt_object",
    "Knife" = "knife;knife",
    "Knife" = "knife;replica",
    "Knife" = "knife;vehicle",
    "Other" = "other",
    "Knife" = "other;blunt_object;knife",
    "Gun" = "other;gun",
    "Knife" = "other;knife",
    "Other" = "replica",
    "Knife" = "replica;knife",
    "Other" = "replica;vehicle",
    "Unarmed" = "unarmed",
    "Unknown" = "undetermined",
    "Unknown" = "unknown",
    "Vehicle" = "vehicle",
    "Gun" = "vehicle;gun",
    "Knife" = "vehicle;knife",
    "Knife" = "vehicle;knife;other"
  )


## add percentage column for labeling
armedinfo <- armedinfo %>% dplyr::mutate(perc = scales::percent(n / sum(n), accuracy = .1, trim = FALSE)) %>% drop_na()


## Plot it
p1 <- ggplot(data = armedinfo, aes(x = n, y = armed_with_rec)) +
  geom_barh(stat = "identity", aes(fill = n)) +
  theme_minimal(base_size = 13) +
  theme(legend.position = "none") +
  scale_x_continuous(expand = c(0, 0)) +
  scale_fill_gradient(low = "royalblue3", high = "navyblue") +
  labs(
    y = NULL,
    x = "Number of deaths",
    title = "Top Suspect Weapon Types in Police Fatal Shootings",
    subtitle = "2015-2024",
    caption = "Ian T. Adams, based on Washington Post data"
  ) +
  geom_text(aes(label = perc),
            hjust = -.25,
            size = 4, fontface = "bold", family = "Fira Sans")

p1 + coord_cartesian(xlim = c(0, 6300), clip = "off")

## Save it
ggsave(filename = here("Output", "top15_suspect-weapon-type.jpg"))







``` 

### Was the Person Killed Fleeing?

```{r}
df_all$flee_status <- df_all$flee_status %>%
  fct_explicit_na("Unknown")

ggplot(data = df, aes(y = flee_status)) + 
        geom_barh(aes(fill = ..count..)) +
        theme_minimal(base_size = 13) +
        theme(legend.position = "none") +
        scale_x_continuous(expand=c(0,0)) +
        scale_fill_gradient(low = "royalblue3", high = "navyblue") +
        labs(y = NULL, x = "Number of deaths")
```
