==> Scripts/03-better stacked plots.Rmd <==
---
title: "03-Better Stacked"
author: "Ian Adams"
date: "4/10/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = FALSE,
	message = FALSE,
	warning = FALSE
)

# General need
library(tidyverse)
library(readr)
library(here)
library(skimr)
library(ggplot2)
library(lubridate)
library(dplyr)

# Specific need
library(showtext)

```

## Import

Don't forget to pull from the git before start (in the "01-Import.Rmd" file), so as to get fresh data from WAPO.

```{r remedy001, message=FALSE, warning=FALSE, include=FALSE}

here::here("WAPO data-police-shootings")

df <- read_rds(here("Data", "Clean", "WAPO_clean.rds"))

df$date <- ymd(df$date) # format as lubridate var
df$year <- year(df$date) # extract year and create variable year
df$month <- month(df$date, label = TRUE, abbr = TRUE) # extract month and create variable month
df$day <- wday(df$date, label = TRUE, abbr = TRUE) # extract day of week, create variable day


## Change N/A to unknown category
df$armed_bucket <- df$armed_bucket %>%
  fct_explicit_na("unknown")

## Reordering df$armed_bucket
df$armed_bucket <- df$armed_bucket %>%
  fct_relevel(
    "gun", "gun like", "sharp", "blunt", "explosive/chemical",
    "vehicle", "taser", "strange", "unarmed", "unknown"
  )



## Recoding df$armed_bucket into df$armed_bucket_rec
df$armed_small <- df$armed_bucket %>%
  fct_recode(
    "armed" = "strange",
    "armed" = "gun like",
    "armed" = "sharp",
    "armed" = "blunt",
    "armed" = "strange object",
    "armed" = "carjack",
    "armed" = "explosive/chemical",
    "armed" = "gun",
    "armed" = "vehicle",
    "armed" = "taser"
  )





```

## Prelim settings

```{r remedy002}

font_add_google("Fira Sans", "firasans")
showtext_auto()

theme_customs <- theme(
  text = element_text(family = 'firasans', size = 16),
  plot.title.position = 'plot',
  plot.title = element_text(
    face = 'bold', 
    colour = thematic::okabe_ito(8)[6],
    margin = margin(t = 2, r = 0, b = 7, l = 0, unit = "mm")
  ),
)

theme_set(theme_minimal() + theme_customs)


```


## Basic plot

```{r remedy003}

df %>% 
  ggplot(aes(x = year, fill = armed_bucket)) +
  geom_bar()

```

## Create groupings

```{r remedy004}


# Group classes into three groups (to reduce colors to 3)
dat3 <- df %>% 
  mutate(
    year = factor(year),
    class_group = case_when(
      armed_bucket %in% c('gun', 'gun like', 'sharp', 'blunt') ~ "Traditional Threat",
      armed_bucket %in% c('strange', 'taser', 'explosive/chemical', 'vehicle') ~ "Rare Threat",
      armed_bucket %in% 'unarmed' ~ "Unarmed",
      armed_bucket == 'unknown' ~ "Unknown"
    )
  )

## Group 4
# dat4 <- df %>% 
#   mutate(
#     year = factor(year),
#     class_group = case_when(
#       armed_bucket %in% c('gun', 'gun like') ~ "Gun/Gun-like",
#       armed_bucket %in% c('sharp', 'blunt') ~ "Sharp/Blunt",
#       armed_bucket %in% c('strange', 'taser', 'explosive/chemical', 'vehicle') ~ "Other Objects",
#       armed_bucket == 'unarmed' ~ "Unarmed",
#       T ~ "Unknowns"
#     )
#   )
# 
# # Group five
# dat5 <- df %>%
#   mutate(
#     year = factor(year),
#     class_group = case_when(
#       armed_bucket %in% c('gun', 'gun like') ~ "Guns",
#       armed_bucket == c('sharp', 'blunt') ~ "Sharps & Blunts",
#       armed_bucket == c(NA, 'unknown') ~ "Unknowns",
#       armed_bucket == c('vehicle','taser','explosive/chemical') ~ "High Threat",
#       armed_bucket == 'unarmed' ~ "Unarmed",
#       T ~ "Strange"
#     )
#   )
# 
# shades_plt <- dat3 %>% 
#   ggplot(aes(x = year, fill = class_group, alpha = armed_bucket)) +
#   geom_bar() +
#   labs(
#     x = 'Year',
#     y = 'Counts',
#     alpha = 'Class',
#     title = 'Police Shootings by Year \nand Suspect Weapon Type'
#   )
# shades_plt 


```


## Color controls

```{r remedy005}

# Color-blind safe colors
# colors <-  thematic::okabe_ito(4)

colorsb <- c("#0072b2","#d55e00",  "#009e73",  "#f0e442")
# Possible levels of transparency (one for each class)
alpha_max <- 1
alpha_min <- 0.1
alpha_vals <- c(
  seq(alpha_max, alpha_min, length.out = 5), 
  seq(alpha_min, alpha_max, length.out = 6)[-1]
)
alpha_vals
## [1] 1.0 0.9 0.8 0.7 0.8 0.9 1.0

# Tweak previous plot
# shades_plt <- shades_plt +
#   scale_fill_manual(values = colors) +
#   scale_alpha_manual(values = alpha_vals)
# shades_plt


dat3 %>% 
  ggplot(aes(x = year, fill = class_group, alpha = armed_bucket)) +
  geom_bar(col = 'white') + # Add lines for distinction
  scale_fill_manual(values = colorsb) +
  scale_alpha_manual(values = alpha_vals) +
  guides(
    #fill = guide_none(),
    alpha = guide_legend(override.aes = list(fill = colorsb[c(2, 2, 2, 2, 1, 1, 1, 1, 3, 4)]))
  ) +
  labs(
    x = 'Year',
    y = 'Counts',
    alpha = 'Type',
    title = 'Fatal Police Shootings by Year and Suspect Weapon Type',
    fill = 'Category'
  ) 

## These auto save, but the defaults are bad, so export manually

ggsave(filename = here("Output", "year-type-group.pdf"))
ggsave(filename = here("Output", "year-type-group.jpg"))
ggsave(filename = here("Output", "year-type-group.png"), bg="white")

```

## Race

```{r remedy006}
library(ggplot2)
library(gghighlight)
library(dplyr)
library(forcats)
library(lubridate)

# Get the current date
current_date <- format(today(), "%m/%d/%Y")

colorsb <- c("#0072b2","#d55e00", "#009e73", "#f0e442", "#56b4e9", "#cc79a7", "#999999", "#E69F00") # extended color palette

## Recoding NA to undetermind
df$race <- df$race %>%
  fct_explicit_na("Undetermined")

# Calculate the percentage of undetermined race for each year
df_percent <- df %>%
  group_by(year, race) %>%
  summarize(count = n()) %>%
  mutate(total = sum(count),
         percent = count / total * 100) %>%
  filter(race == "Undetermined")

# Create the bar plot
p <- df %>% ggplot(aes(x=year, fill=race)) +
  geom_bar(col = 'white') +
  scale_fill_manual(values = colorsb) +
  labs(
    x = 'Year',
    y = 'Counts',
    title = 'Fatal Police Shootings by Year and Race',
    subtitle = paste0('WaPo Fatal Shootings Data as of ', current_date),
    fill = 'Race'
  ) + scale_x_continuous(breaks = c(2015,2016,2017,2018,2019,2020,2021,2022,2023)) #+
  gghighlight(race == 'Undetermined', use_direct_label = F)

# Add percentage labels for undetermined race
p + geom_text(data = df_percent, aes(x = year, y = count, label = paste0(round(percent, 1), "%")), position = position_stack(vjust = 0.5), color = "black")

# Save the plot
ggsave(filename = here("Output", "undetermined_highlight.png"), bg="white")


```

## Race total 

2015-2020 only! 2021 unidentified race is too overwhelmingly bad right now

```{r remedy007}
#Make "NA" into an explicit factor level
df2 <- df %>% filter(year < 2021) ## Ignore 2021,2022

df2$race_rec <- df2$race_rec %>%
  fct_explicit_na("Unknown")

df2$race_rec <- factor(df2$race_rec,
                         levels = names(sort(table(df2$race_rec), 
                                             decreasing = TRUE)))

ggplot(df2, aes(x = race_rec)) + 
  geom_bar(fill=ggpubfigs::friendly_pal("bright_seven")) +
  geom_text(aes(label = sprintf("%0.2f", round(..count..)/sum(..count..),digits=2)), stat = "count", vjust = -.5, colour = "black") +
  labs(
    x = 'Race',
    y = 'Counts',
    title = 'Fatal Police Shootings by Race',
    fill = 'Race',
    caption = "Counts (axis) and proportions (label)"
  )

```

## Better line plots

```{r remedy007}

library(gghighlight)

df <- df %>% mutate(row_number= 1:n())

yearly_counts <- df %>%
  count(year, armed_bucket)

yearly_counts %>% ggplot(aes(x = year, y = n, color = armed_bucket)) +
    geom_line()

## Or as one

# Color-blind safe colors
colors <- thematic::okabe_ito(11)[-3]

df %>% filter(year < 2021) %>% na.omit() %>%
  count(year,armed_bucket) %>%
  ggplot(aes(x = year, y = n, color = armed_bucket)) +
    geom_line(size = 1) +
  scale_color_manual(values = colors) +
  #coord_cartesian(xlim = c(15, 81), expand = F) +
  scale_y_continuous(minor_breaks = NULL) +
  #scale_x_date(date_breaks = "4 months", date_labels = "%B\n%y") +
  labs(x = 'Time', y = 'Count', col = 'Weapon Type', title = "In fatal police shootings...") +
  gghighlight(armed_bucket == 'gun' | armed_bucket=='gun like' | armed_bucket=="sharp", use_direct_label = T) +
  annotate(
    'text',
    x = 2015,
    y = 575,
    label = '...most suspects were armed with a gun or bladed object',
    hjust = 0,
    vjust = 0,
    family = 'firasans',
    size = 5
  )

df$Month_Yr <- format(as.Date(df$date), "%Y-%m")
#df$Month_Yr <- as.Date(as.yearmon(df$Month_Yr))
 

ggplot(df) +
  aes(x = date, y = row_number) +
  geom_line(size = 0.5, colour = "#112446") +
  scale_x_date(date_breaks = "4 months", date_labels = "%B\n%y") +
  theme_minimal()

```
